# api/openapi.yaml
openapi: 3.0.3
info:
  title: Zyad Cloud API
  version: 0.1.0
paths:
...
  /api/v1/admin/landing-pages/{id}/revisions:
    get:
      summary: List landing page revisions
      operationId: listLandingPageRevisions
      tags:
        - Landing Revisions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Landing page revisions retrieved successfully
  /api/v1/admin/landing-pages/{id}/revisions/compare:
    get:
      summary: Compare landing page revisions
      operationId: compareLandingPageRevisions
      tags:
        - Landing Revisions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "501":
          description: Revision compare is not implemented yet
...
  /api/v1/admin/landing-pages/{id}/sections:
    get:
      summary: List landing page sections
      operationId: listLandingPageSections
      tags:
        - Landing Sections
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Sections retrieved successfully
    post:
      summary: Create landing page section
      operationId: createLandingPageSection
      tags:
        - Landing Sections
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Section created successfully
  /api/v1/admin/landing-pages/{id}/sections/reorder:
    put:
      summary: Reorder landing page sections
      operationId: reorderLandingPageSections
      tags:
        - Landing Sections
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Sections reordered successfully
...
  /api/v1/admin/landing-pages/{id}/sections/{sectionId}:
    patch:
      summary: Update landing page section
      operationId: updateLandingPageSection
      tags:
        - Landing Sections
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Section updated successfully
    delete:
      summary: Delete landing page section
      operationId: deleteLandingPageSection
      tags:
        - Landing Sections
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Section deleted successfully
...
  /api/v1/admin/landing-submissions:
    get:
      summary: List landing submissions
      operationId: listLandingSubmissions
      tags:
        - Landing Submissions
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Submissions retrieved successfully
  /api/v1/admin/landing-submissions/export:
    get:
      summary: Export landing submissions
      operationId: exportLandingSubmissions
      tags:
        - Landing Submissions
      security:
        - BearerAuth: []
      responses:
        "501":
          description: Export functionality is not yet implemented
  /api/v1/admin/landing-submissions/{id}:
    get:
      summary: Get landing submission
      operationId: getLandingSubmission
      tags:
        - Landing Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Submission retrieved successfully
    delete:
      summary: Delete landing submission
      operationId: deleteLandingSubmission
      tags:
        - Landing Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Submission deleted successfully
  /api/v1/admin/landing-submissions/{id}/status:
    patch:
      summary: Update landing submission status
      operationId: updateLandingSubmissionStatus
      tags:
        - Landing Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Submission status updated successfully
  /api/v1/admin/landing-submissions/{id}/notes:
    post:
      summary: Add landing submission note
      operationId: addLandingSubmissionNote
      tags:
        - Landing Submissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Submission note added successfully
...
  /api/v1/public/landing/forms/{formKey}/submissions:
    post:
      summary: Submit public landing form
      operationId: submitPublicLandingForm
      tags:
        - Public Landing
      parameters:
        - name: formKey
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Public landing form submitted successfully
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...